requirements.txt
Flask
python-docx
python-dotenv
huggingface-hub
gunicorn
groq